pyproject.toml
[project]
name = "scenario-analysis"
version = "0.1.0"
description = ""
authors = [
    {name = "Kevin Furtlehner",email = "furtikevin@gmail.com"}
]
packages = [
    { include = "scenario_analysis", from = "src" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "lxml (>=6.0.2,<7.0.0)",
    "openai (>=2.14.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
